Create ODM formated dataframe

The function ODMcreate can be used to help in formating new data. ODMcreate takes a vector of dates and data and outputs a dataframe formated for use with the functions in this package. Site ID, VariableID, and MethodID need to already exist in their respective tables in the ODM.

ODMcreate(LocalDateTime, DataValue, UTCOffset = -5, SiteID, VariableID,
  QualifierID = NA, MethodID, SourceID = 1, QCLevelID = 0)

Arguments

LocalDateTime
local date time for each record

DataValue
matching data value for each date time record

UTCOffset
difference in hours from Coordinated Universal Time (UTC)

SiteID
corresponds to an existing Sites record in the Sites table

VariableID
corresponds to an existing record in the Variables table

QualifierID
corresponds to an existing record in the Qualifiers table

MethodID
corresponds to an existing record in the Methods table

SourceID
corresponds to an existing record in the Sources table

QCLevelID
corresponds to a record in the QualityControlLevels table

Details

The vectors are given with the arguments for SiteID, VariableID and MethodID. The output is a properly formated dataframe suitable for further use in workflows involving this package.

Examples

# NOT RUN {
# dates
date = seq(as.Date("2010/1/1"), as.Date("2011/1/1"), "days")
# dummy data values
value = 1:366

# creation of dataframe that meets the requirements of the ODM.
tmp = ODMcreate(LocalDateTime = date, DataValue = value, SiteID = 1,
 VariableID = 1, MethodID = 9)
# }